+++ CHANGELOG.md
@@ -4,6 +4,21 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
+## [Unreleased]
+
+### Added 
+
+- Add `SitterProfileViewModel` - [#164](https://github.com/ios-course/ironfoudation-team-project/pull/164)
+- Integrate `NetworkService` package into the `WoofApp` - [#188](https://github.com/ios-course/ironfoudation-team-project/pull/188)
+- Add the ability to create and save a new owner or sitter when choosing a role on the `LoginView` screen - [#192](https://github.com/ios-course/ironfoudation-team-project/pull/192)
+- Add `SitterProfileView` to display and edit personal information about the sitter - [#190](https://github.com/ios-course/ironfoudation-team-project/pull/190)
+- Add `Obfuscator` enum with `reveal(_:_:)` method - [#196](https://github.com/ios-course/ironfoudation-team-project/pull/196) 
+- Add bash script `Obfuscator.sh` - [#197](https://github.com/ios-course/ironfoudation-team-project/pull/197)
+
+### Changed
+
+ - Update PR template [#194](https://github.com/ios-course/ironfoudation-team-project/pull/194)
+
 ## [0.3.1] - 21.08.2023
 
 ### Changed